Update LCC test.
